requirements.txt
ipykernel
setuptools
pinecone[grpc]==6.0.2
openai
scikit-learn
pillow
opencv-python
python-dotenv==1.1.1
pytest
fastapi
uvicorn
pydantic
streamlit
soundfile==0.13.1
vosk==0.3.44
edge_tts==7.2.3
mlflow==3.5.1
python-multipart==0.0.20